Add files via upload
xploratory Data Analysis (EDA) is an approach to analyzing data sets to summarize their main characteristics, often with visual methods. EDA is for seeing what the data can tell us beyond the formal modelling or hypothesis testing task.
Most EDA techniques are graphical in nature with a few quantitative techniques. The reason for the heavy reliance on graphics is that by its very nature the main role of EDA is to open-mindedly explore, and graphics gives the analysts unparalleled power to do so, enticing the data to reveal its structural secrets, and being always ready to gain some new, often unsuspected, insight into the data. In combination with the natural pattern-recognition capabilities that we all possess, graphics provides, of course, unparalleled power to carry this out. 
The particular graphical techniques employed in EDA are often quite simple, consisting of various techniques of: 
1.Plotting the raw data (such as data traces, histograms, bihistograms, probability plots, lag plots, block plots, and Youden plots.

2.Plotting simple statistics such as mean plots, standard deviation plots, box plots, and main effects plots of the raw data.

3.Positioning such plots so as to maximize our natural pattern-recognition abilities, such as using multiple plots per page.